Set git pager to exit if output is one screen full
Git pager is set to `less -FMR`; where -F flag causes less to
automatically exit if the entire output can be displayed on the first
screen, -M flag is for more verbose and -R flag is for showing colors.
The output will not be cleared from the screen even if it is more than
one screen full. That will require -+X flag, which will not work well
with -F flag, so, it should be changed to -+F flag. But I like that I
can see latest screen of output after quitting from the pager.