Fix volume blocklet in i3blocks
Volume block was set to update when received a signal. But when a
headphone was plugged in/out the block doesn't receive any signal and
doesn't update. One way is to add interval of 1 sec instead of once but
that is not encouraged. Instead I swapped the volume script with
volume-pulseaudio from i3blocks-contrib repo by the advice from kb100.