Fix vim-tmux-navigation
Commit fecac58666e1d3b149a1b425043b86d0824ac92f had checked for zoomed
tmux panes before switching. That is if the current pane was zoomed,
nothing would be done. But that included panes containing vim too, and
if the current tmux pane was zoomed and the vim contained more than one
split, the keybindings to switch panes would do nothing in vim.

This commit fixes that by sending keys directly to the tmux pane if the
pane is zoomed, so that irrespective of vim being in that pane, that
pane would receive the pressed keys, no tmux pane switching would be
done. And if the pane is not zoomed, the previous behavior is adapted.