Update fzf configurations
The check for silversearcher would output error if `ag` was not found.
This has been fixed by redirecting the std error to /dev/null.
Unfortunately, then the result would output the confirmation of the
presence of `ag`. So both the std output and std error were redirected
to /dev/null.