Change action of tab button in insert-mode in vim
Use tabs for indentation and spaces for alignment. Tab key will insert
tabs at the beginning of a line and intelligent amount of spaces
elsewhere. There most definitely is a lot of bugs and corner cases. The
one I've already encountered is, when you have characters in current
line and you press <Tab> at the beginning, instead of inserting a tab
character, it inserts spaces in between the next word. But it doesn't
bother me, because if I want to change the indent level, I can do `>` or
`<` in normal-mode and `<C-t>` or `<C-d>` in insert-mode.