Add function for retrieving file names
The insert mode file search, `:h i_CTRL-X_CTRL-F`, depends on the
'wildignore' setting. So, if I set wildignore to ignore all jpg files, I
won't be able to insert the filename of a jpg file with `i_CTRL-X_CTRL-F`

But this function, MyCompleteFileName(), retrieves the filename without
paying heed to 'wildignore' setting. After binding it to <C-F> in insert
mode, it would solve the aforementioned problem.