Change only the local statusline in markdown
For markdown filetype, the statusline is modified to show the word count
of the buffer. If the statusline is set with `set` instead of `setlocal`
then the statusline is modified for all windows. I don't want that. I
want only the window containing markdown buffer to have a modified
statusline.

Using `setlocal` instead of `local` takes care of that for me.