Fix: neovim cursor in WSL
Neovim cursor-shapes for different modes doesn't work. So I had to
disable it. I'm quite used to it now. But for whatever reason, in WSL,
the `se gcr=` stopped working for a moment. I started commenting out and
uncommenting to see what config was causing the problem. Found nothing.
Everything started working after a while. Makes you believe in god.