Fix powershell functions
Powershell aliases aren't as versatile as bash's. In PS, the aliases can
take only the cmdlet as the value, not the arguments. So you need to
define a function with all the arguments for the base cmdlet and then
create an alias to that new function.

Previously I forgot to add the arguments in that function. So the args
for the aliases weren't passed to the function. It's fixed now.