Fix powershell prompt
https://github.com/MicrosoftDocs/terminal/issues/425

In Windows Terminal, if you want to open a new tab or split in $pwd,
you'll have to add OSC9;9 string after the prompt. But adding that
messed with the output when using grep with pipe.

It turns out, I was using an object as a string. Fixing that seems to
have fixed the bug.