VIM: fix guicursor setting to nil
For some reason the following lines disrupts the cursor in a way that if
guicursor is set later to nil it's set to block and not what it was
before. Must be some weird bug. I don't even know why I wrote this part.
I don't use neovim inside msys.

    elseif system('uname') =~? "msys"
        let g:env = 'MSYS2'