sig: Remove sourcehut link
It's gonna be a paid service. Although I've kinda fallen in love with
its eccentric (but sensible) philosophy, paying for a hosting service
month after month doesn't look feasible. I'll try to adopt to sourcehut
and be a paid member, but until that's happened already, no sense in
advertising an empty corner I haven't set up home in.