vim: use noignorecase by default
Problem: `ignorecase` is followed even for `gd`, which matches keywords
with wrong case.
Solution: Turn off `ignorecase` by default, and set it on for each /,?,
and :substitute.
Remarks: A better way maybe would've been to map `gd` to a custom
command which behaves in a `noignorecase` manner, but it's a complex
function with no exposed vimL implementation. Digging through C source
code is a task for another time.