Say only "Enan" in personal & work email signatures
Enan is an uncommon enough name that it can be used without full name for
unique identification.  Most of my work emails are discussion-centric,
it not sensible to clutter the thread with a banner in signature.  If I
have to send a formal email, which should be few and far between, I'll
hand-craft the signature.