font: replace OCR-A Extended with John Sauter's OCR-A
The OCR-A Extended font that comes with Windows is rendered with extra
gap between each pair of characters in Simple Terminal.  The version
developed by John Sauter doesn't cause that problem.  I don't what the
cause is.

The only difference I could find between the two implementations of
OCR-A is that in the Windows version, the tilde is vertically aligned in
the center.  I have come to like the eccentric position (top-aligned) of
tilde in John Sauter's version.

Ref: http://luc.devroye.org/OCRA.zip